backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-decouple==3.8
alembic==1.12.1
pydantic[email]==2.6.0
pydantic-settings==2.2.1
pytest==7.4.3
httpx==0.25.2

# Nouvelles dépendances pour les fonctionnalités avancées
slowapi==0.1.9
websockets==12.0
sse-starlette==1.8.2
prometheus-client==0.19.0
python-multipart==0.0.6
magic==0.4.27
loguru==0.7.2
elasticsearch==8.11.0
redis==5.0.1
